--- src/main/java/io/polyrepo/analyser/repository/IQueryRepository.java
@@ -1,13 +1,0 @@
-package io.polyrepo.analyser.repository;
-
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Map;
-
-
-@Repository
-public interface IQueryRepository {
-
-    List<Map<String, Object>> getStoredQueries(int userId) throws IndexOutOfBoundsException;
-}
